[webui-testsuite] speedup the spider test
In not playing ping pong with the browser, but just doing all parsing and
validation ourselve parsing the page_source, we save roughly 50% of the test
runtime - which is insane

+++ src/webui-testsuite/run_acceptance_tests.rb
@@ -162,10 +162,12 @@ while true
           # OK
         else
           puts "Webui did not response nicely"
-          Process.kill "INT", webui_out.pid
-          webui_out.close
+	  if webui_out
+            Process.kill "INT", webui_out.pid
+            webui_out.close
+	  end
           webui_out = nil
-          frontend.join
+          frontend.join if frontend
 	  puts lines.join()
           exit 1
       end
